Add custom wording and ability to hide close box.
Some organizations want to use our own wording in the header
and the body. This PR enables that.

Also, if we want to show a notice that opens this modal in a
new blank tab, we will want to disable closing (because closing the
modal on a blank tab just shows a blank page). This PR also adds
that capability.

The README has been updated to explain how to take advantage of both
these features.